+++ Src/Object/Character/Enemy/EnemyOnHit.cpp
@@ -71,7 +71,7 @@ void EnemyOnHit::CollNormalAttack(const std::weak_ptr<Collider> _hitCol)
 	cardDif == 0 ? addDam = 0 : addDam = MAX_ADD_DAMAGE - cardDif;
 
 	//ダメージ計算
-	int damage = static_cast<float>(atkPoint) + addDam;
+	int damage = static_cast<int>(atkPoint) + addDam;
 	charaObj_.Damage(damage);
 
 	action_.ChangeAction(ActionController::ACTION_TYPE::REACT);
